+++ functions/Get-ChocoOutdatedPackages.ps1
@@ -1,3 +1,5 @@
+
+function Get-ChocoOutdatedPackages {
 <#
 .SYNOPSIS
     Find outdated packages from a local machine
@@ -13,7 +15,6 @@ curl                 7.64.0         7.64.1        false
 GoogleChrome         73.0.3683.103  74.0.3729.131 false
 
 #>
-function Get-ChocoOutdatedPackages {
     [CmdletBinding()]
     param(
     )